requirements.txt
# pyenv virtualenv context-aware-resizing
numpy
scipy
matplotlib
Pillow
numba